Handle StandardError exceptions raised by the strategy#process.
We realized that a strategy#process might raise a StandardError exception that
wouldn't allow the framework to log the failed item. So, we added the rescue
for that exception, log the failed item and then raise the exception.
This is very helpful, because, now we have the failed item in the log.